linter: remove prefer-big-int
