+++ crates/squawk_linter/src/rules/mod.rs
@@ -14,7 +14,6 @@ pub(crate) mod ban_drop_table;
 pub(crate) mod changing_column_type;
 pub(crate) mod constraint_missing_not_valid;
 pub(crate) mod disallow_unique_constraint;
-pub(crate) mod prefer_big_int;
 pub(crate) mod prefer_bigint_over_int;
 pub(crate) mod prefer_bigint_over_smallint;
 pub(crate) mod prefer_identity;
@@ -43,7 +42,6 @@ pub(crate) use ban_drop_table::ban_drop_table;
 pub(crate) use changing_column_type::changing_column_type;
 pub(crate) use constraint_missing_not_valid::constraint_missing_not_valid;
 pub(crate) use disallow_unique_constraint::disallow_unique_constraint;
-pub(crate) use prefer_big_int::prefer_big_int;
 pub(crate) use prefer_bigint_over_int::prefer_bigint_over_int;
 pub(crate) use prefer_bigint_over_smallint::prefer_bigint_over_smallint;
 pub(crate) use prefer_identity::prefer_identity;
