--- docs/docs/prefer-big-int.md
@@ -1,12 +1,0 @@
----
-id: prefer-big-int
-title: prefer-big-int
----
-
-## deprecated
-
-:::caution Deprecated
-This rule has been deprecated in favor of ["prefer-bigint-over-int"](./prefer-bigint-over-int.md) and ["prefer-bigint-over-smallint"](./prefer-bigint-over-smallint.md).
-:::
-
-This rule has been split into ["prefer-bigint-over-int"](./prefer-bigint-over-int.md) and ["prefer-bigint-over-smallint"](./prefer-bigint-over-smallint.md) to allow more granular linting against integer fields. Please use this new rules as this rule will be removed in a future release.
